+++ __tests__/components/GuideCard.test.tsx
@@ -19,7 +19,7 @@ import {
   FeedbackStatus,
   GradesGivenStatus,
   ReturnStatus,
-} from "../../app/guides/types";
+} from "../../types/guideTypes";
 import exp from "constants";
 
 const mockGuide = {
